remove reads that are both discordant at one integration, and chimeric at the same or a different integration.
A read pair that is discordant about an integration, but with one read that crosses it as chimeric currently can't be detected by the pipeline. In order to be detected by the pipeline as a discordant read-pair, read1 must be mapped and read2 unmapped in one alignment, and vice versa for the other alignment.  In order to be considered 'mapped', a read must have at most some nubmer of bases unmapped (default 20), and in order to be considered 'unmapped', a read can only have at most the same number of bases mapped.  Therefore, reads that cross an integration (are chimeric) cannot be involved in a discordant read pair.  The reason for the threshold of mapped/unmapped bases is mostly to avoid supurious alignments coming out as discordant read-pairs.

This arrangement creates problems when scoring reads which are both chimeric and in discordant pairs.  Since the discordant pair can't be detected by the pipeline, they are scored as 'false negatives', despite the fact that this is the desired behaviour of the pipeline.  So, if a read is involved in both a chimeric junction and a discordant read pair, ignore the fact that it is discordant (since it won't be detected).  This removal is justified in two different cases:

In the case where the chimeric read and the discordant pair are in reference to ths same junction, there is no loss of information from removing the discordant pair, since we have already detected the integration with the chimeric read.

There is also the case where the chimeric read is at one jucntion at the discordant pair is at a different junction originating from a different, nearby integration. In this case, there is some loss of information (namely, the presence of the second integration).  However, we assume these events could be exceedingly rare in real data, since for this situation to arise, two integrations would have to occur next to each other in the same cell. If the virus/vector has no particular strong biases towards particular locations in the genome, then these events are every unlikely to occur (but they may occur in viruses with strong integration location preferences).  But these events can occur relativley commonly in simulated data, since in the simulation all integrations occur in the same 'cell'. Since this is an artifact of the way the simluations are performed, and these events should be rare in real data, ignore the discordant pair in this case.